requirements.txt
keras==3.3.3
numpy==1.24.3
opencv_python==4.9.0.80
Pillow==10.0.1
streamlit==1.34.0
toml==0.10.2
torch==2.3.0
torchvision==0.18.0
